+++ src/Modules/Legal/Legal.js
@@ -1,16 +1,9 @@
 import '@/core'
 import '@/globals'
-import i18n from '@/helper/i18n'
+import LegalPage from './components/LegalPage.vue'
+import { vueApply, vueRegister } from '@/vue'
 
-const noticeId = '#legal_form_privacyNoticeAcknowledged'
-const form = document.querySelector('form[name="legal_form"]')
-const doNotAgree = '0'
-
-form.addEventListener('submit', function (event) {
-  const noticeElement = document.querySelector(noticeId)
-  if (noticeElement && noticeElement.value === doNotAgree) {
-    if (!confirm(i18n('legal.are_you_sure_to_downgrade'))) {
-      event.preventDefault()
-    }
-  }
+vueRegister({
+  LegalPage,
 })
+vueApply('#legal-page')